@@ -5,7 +5,7 @@ description: A highly customizable typeahead (autocomplete) text input field for
 homepage: https://github.com/AbdulRahmanAlHamali/flutter_typeahead
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.17.0"
 
 
